requirements.txt
# Flask Core
Flask==3.0.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database - SQL Server
pyodbc>=4.0.39
SQLAlchemy==2.0.31
alembic==1.13.1

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
numpy==2.1.3

# API & Security
marshmallow==3.20.2
bcrypt==4.1.2
python-dotenv==1.0.1

# Utilities
qrcode==8.2.0
Pillow==11.3.0

# Monitoring & Logging (optional)
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-flask==1.3.0